seznam/pyproject.toml
[project]
name = "seznam"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "grpcio>=1.74.0",
    "protobuf>=6.32.1",
    "pydantic>=2.11.9",
    "szn-proto-asistent>=0.2.8211042",
    "tinygent",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["seznam"]

[[tool.uv.index]]
url = "https://pypi.repo.ops.iszn.cz/local/simple"

[[tool.uv.index]]
url = "https://pypi.repo.ops.iszn.cz/simple"
default = true

[tool.uv.sources]
tinygent = { path = "../" , editable = true }
